backend_compiled/backend/airflow/requirements.txt
# PDF + DB
PyPDF2
pdfplumber
requests
beautifulsoup4
PyYAML
SQLAlchemy
# any others your scripts need…
